Initial version of BLE-PKAP example
-- Partial example implemenation of the BLE Public Key Authenticated Pairing
(BLE-PKAP) protocol.  This code includes both a initiator example (written in
python, for use on Linux/Bluez) and a responder example (written in C++ with
the nRF5 SDK for use on a Nordic nRF52840 development board).  This version of
the code implements the overall protocol flow, but includes mock code for the
public key authentication logic.